requirements.txt
# Core Django
Django>=5.0,<5.1
django-ninja>=1.0.0
django-cors-headers>=4.3.0

# HTTP Client
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-asyncio>=0.21.0
pytest-rerunfailures>=12.0

# Development
python-dotenv>=1.0.0

# Production
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Database (PostgreSQL support)
psycopg2-binary>=2.9.0

# Redis
redis>=5.0.0

# Celery
celery>=5.3.0

# Monitoring
sentry-sdk[django]>=1.38.0

# Static files
whitenoise>=6.6.0

# Security
cryptography>=41.0.0

# Utilities
Pillow>=10.0.0
python-dateutil>=2.8.0